Rename pure run package to execution
