+++ internal/pure/execution/failure.go
@@ -1,4 +1,4 @@
-package run
+package execution
 
 import "github.com/becker63/searchbench-go/internal/pure/domain"
 
@@ -15,14 +15,14 @@ const (
 	FailureReport FailureStage = "report"
 )
 
-// RunFailure records the report-facing failed path for one task/system run.
+// RunFailure records the report-facing failed path for one match/system run.
 //
 // ExecutedRun means execution succeeded. score.ScoredRun means scoring
 // succeeded. RunFailure is the separate artifact used when one of those steps
 // could not complete.
 type RunFailure struct {
 	RunID   domain.RunID    `json:"run_id"`
-	TaskID  domain.TaskID   `json:"task_id"`
+	MatchID domain.MatchID  `json:"match_id"`
 	System  domain.SystemID `json:"system"`
 	Stage   FailureStage    `json:"stage"`
 	Message string          `json:"message"`
@@ -32,7 +32,7 @@ type RunFailure struct {
 func NewFailure(spec Spec, stage FailureStage, message string) RunFailure {
 	return RunFailure{
 		RunID:   spec.ID,
-		TaskID:  spec.Task.ID,
+		MatchID: spec.Match.ID,
 		System:  spec.System.ID,
 		Stage:   stage,
 		Message: message,
