+++ internal/pure/execution/phases.go
@@ -1,4 +1,4 @@
-package run
+package execution
 
 // Phase is a sealed marker interface for run lifecycle phases.
 //
